Reorganize analysis code
