+++ virtool/jobs/analysis.py
@@ -39,320 +39,6 @@ LIST_PROJECTION = [
 ]
 
 
-async def new(db, settings, manager, sample_id, user_id, algorithm):
-    """
-    Creates a new analysis. Ensures that a valid subtraction host was the submitted. Configures read and write
-    permissions on the sample document and assigns it a creator username based on the requesting connection.
-
-    """
-    # Get the current id and version of the species index currently being used for analysis.
-    index_id, index_version = await virtool.db.indexes.get_current_index(db)
-
-    sample = await db.samples.find_one(sample_id, ["name"])
-
-    analysis_id = await virtool.utils.get_new_id(db.analyses)
-
-    job_id = await virtool.utils.get_new_id(db.jobs)
-
-    document = {
-        "_id": analysis_id,
-        "ready": False,
-        "created_at": virtool.utils.timestamp(),
-        "job": {
-            "id": job_id
-        },
-        "algorithm": algorithm,
-        "sample": {
-            "id": sample_id
-        },
-        "index": {
-            "id": index_id,
-            "version": index_version
-        },
-        "user": {
-            "id": user_id,
-        }
-    }
-
-    task_args = dict(
-        analysis_id=analysis_id,
-        sample_id=sample_id,
-        sample_name=sample["name"],
-        index_id=index_id
-    )
-
-    if "pathoscope" in algorithm:
-        sequence_species_map = dict()
-        species_dict = dict()
-
-        async for sequence_document in db.sequences.find({}, ["species_id", "isolate_id"]):
-            species_id = sequence_document["species_id"]
-
-            species = species_dict.get(species_id, None)
-
-            if species is None:
-                species = await db.species.find_one(species_id, ["last_indexed_version"])
-
-                try:
-                    last_index_version = species["last_indexed_version"]
-
-                    species_dict[species["_id"]] = {
-                        "id": species["_id"],
-                        "version": last_index_version
-                    }
-
-                    sequence_species_map[sequence_document["_id"]] = species_id
-                except KeyError:
-                    species_dict[species["id"]] = False
-
-            sequence_species_map[sequence_document["_id"]] = species_id
-
-        sequence_species_map = [item for item in sequence_species_map.items()]
-
-        task_args.update({
-            "species_dict": species_dict,
-            "sequence_species_map": sequence_species_map
-        })
-
-    await db.analyses.insert_one(document)
-
-    # Clone the arguments passed from the client and amend the resulting dictionary with the analysis entry
-    # _id. This dictionary will be passed the the new analysis job.
-    await manager.new(
-        document["algorithm"],
-        task_args,
-        user_id,
-        job_id=job_id
-    )
-
-    await virtool.db.samples.recalculate_algorithm_tags(db, sample_id)
-
-    return document
-
-
-def coverage_to_coordinates(coverage_list):
-    previous_depth = coverage_list[0]
-    coordinates = {(0, previous_depth)}
-
-    last = len(coverage_list) - 1
-
-    for i, depth in enumerate(coverage_list):
-        if depth != previous_depth or i == last:
-            coordinates.add((i - 1, previous_depth))
-            coordinates.add((i, depth))
-
-            previous_depth = depth
-
-    coordinates = sorted(list(coordinates), key=lambda x: x[0])
-
-    return coordinates
-
-
-def get_square_distance(p1, p2):
-    """
-    Get the square of the distance between two points.
-
-    """
-    dx = p1[0] - p2[0]
-    dy = p1[1] - p2[1]
-
-    return dx * dx + dy * dy
-
-
-def get_square_segment_distance(p, p1, p2):
-    """
-    Get the square distance between a point and a segment.
-
-    """
-    x, y = p1
-
-    dx = p2[0] - x
-    dy = p2[1] - y
-
-    if dx != 0 or dy != 0:
-        t = ((p[0] - x) * dx + (p[1] - y) * dy) / (dx * dx + dy * dy)
-
-        if t > 1:
-            x, y = p2
-        elif t > 0:
-            x += dx * t
-            y += dy * t
-
-    dx = p[0] - x
-    dy = p[1] - y
-
-    return dx * dx + dy * dy
-
-
-def simplify_radial_distance(points, tolerance):
-    prev_point = points[0]
-    new_points = [prev_point]
-
-    point = None
-
-    for point in points:
-        if get_square_distance(point, prev_point) > tolerance:
-            new_points.append(point)
-            prev_point = point
-
-    if point is not None and prev_point != point:
-        new_points.append(point)
-
-    return new_points
-
-
-def simplify_douglas_peucker(points, tolerance):
-    length = len(points)
-    markers = [0] * length
-
-    first = 0
-    last = length - 1
-
-    first_stack = list()
-    last_stack = list()
-
-    new_points = list()
-
-    markers[first] = 1
-    markers[last] = 1
-
-    while last:
-        max_squared_distance = 0
-
-        for i in range(first, last):
-            squared_distance = get_square_segment_distance(points[i], points[first], points[last])
-
-            if squared_distance > max_squared_distance:
-                index = i
-                max_squared_distance = squared_distance
-
-        if max_squared_distance > tolerance:
-            markers[index] = 1
-
-            first_stack.append(first)
-            last_stack.append(index)
-
-            first_stack.append(index)
-            last_stack.append(last)
-
-        # Can pop an empty array in Javascript, but not Python, so check
-        # the length of the list first
-        try:
-            first = first_stack.pop()
-        except IndexError:
-            first = None
-
-        try:
-            last = last_stack.pop()
-        except IndexError:
-            last = None
-
-    for i, point in enumerate(points):
-        if markers[i]:
-            new_points.append(point)
-
-    return new_points
-
-
-def smooth_coverage_coordinates(coordinates, tolerance=0.01):
-    squared_tolerance = pow(tolerance, 2)
-
-    return simplify_douglas_peucker(simplify_radial_distance(coordinates, squared_tolerance), squared_tolerance)
-
-
-async def format_analysis(db, analysis):
-    if "pathoscope" in analysis["algorithm"]:
-        formatted = dict()
-
-        for hit in analysis["diagnosis"]:
-
-            species_id = hit["species"]["id"]
-            version = hit["species"]["version"]
-
-            species = formatted.get(species_id, None)
-
-            if species is None:
-                # Get the species entry (patched to correct version).
-                _, species_document, _ = await virtool.db.history.patch_to_version(
-                    db,
-                    species_id,
-                    version
-                )
-
-                max_ref_length = 0
-
-                for isolate in species_document["isolates"]:
-                    max_ref_length = max(max_ref_length, max([len(s["sequence"]) for s in isolate["sequences"]]))
-
-                species = {
-                    "id": species_id,
-                    "name": species_document["name"],
-                    "version": species_document["version"],
-                    "abbreviation": species_document["abbreviation"],
-                    "isolates": species_document["isolates"],
-                    "length": max_ref_length
-                }
-
-                formatted[species_id] = species
-
-            for isolate in species["isolates"]:
-                for sequence in isolate["sequences"]:
-                    if sequence["_id"] == hit["id"]:
-                        sequence.update(hit)
-                        sequence["length"] = len(sequence["sequence"])
-
-                        del sequence["species"]
-                        del sequence["species_id"]
-                        del sequence["isolate_id"]
-
-        analysis["diagnosis"] = [formatted[species_id] for species_id in formatted]
-
-        for species in analysis["diagnosis"]:
-            for isolate in list(species["isolates"]):
-                if not any((key in sequence for sequence in isolate["sequences"]) for key in ("pi", "final")):
-                    species["isolates"].remove(isolate)
-                    continue
-
-                for sequence in isolate["sequences"]:
-                    if "final" in sequence:
-                        sequence.update(sequence.pop("final"))
-                        del sequence["initial"]
-                    if "pi" not in sequence:
-                        sequence.update({
-                            "pi": 0,
-                            "reads": 0,
-                            "coverage": 0,
-                            "best": 0,
-                            "length": len(sequence["sequence"])
-                        })
-
-                    sequence["id"] = sequence.pop("_id")
-
-                    if "align" in sequence:
-                        coordinates = coverage_to_coordinates(sequence["align"])
-                        sequence["align"] = smooth_coverage_coordinates(coordinates, tolerance=0.05)
-
-                    del sequence["sequence"]
-
-        return analysis
-
-    if analysis["algorithm"] == "nuvs":
-        for sequence in analysis["results"]:
-            for orf in sequence["orfs"]:
-                for hit in orf["hits"]:
-                    hmm = await db.hmm.find_one({"_id": hit["hit"]}, [
-                        "cluster",
-                        "families",
-                        "names"
-                    ])
-
-                    hit.update(hmm)
-
-        return analysis
-
-    raise ValueError("No algorithm field in document")
-
-
 class Base(virtool.jobs.job.Job):
     """
     A base class for all analysis job objects. Functions include:
